s05/Assignment/pyproject.toml
[project]
name = "dog-breed-classifier"
version = "0.1.0"
description = "Dog Breed Classifier with PyTorch Lightning and Hydra"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "lightning>=2.4.0",
    "loguru>=0.7.3",
    "pandas>=2.0.0",
    "rich>=13.0.0",
    "torchvision>=0.18.0",
    "hydra-core>=1.3.0",
    "hydra-colorlog>=1.2.0",
    "omegaconf>=2.3.0",
    "rootutils>=1.0.7",
    "matplotlib>=3.8.0",
    "pillow>=10.0.0",
    "pytest>=8.0.0",
    "coverage>=7.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "coverage>=7.0.0",
    "black>=24.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::UserWarning",
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/__pycache__/*", "*/tests/*", "src/app.py", "src/utils/logging_config.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "raise FileNotFoundError",
    "logger.debug",
    "logger.info", 
    "logger.warning",
    "logger.error",
    "console.print",
    "plt.show",
    "plt.savefig",
    "plt.close",
    "trainer.fit",
    "trainer.validate",
    "trainer.test",
    "hydra.main",
    "@hydra.main",
]
fail_under = 70
